+++ bin/fetchSwagger.js
@@ -1,9 +1,10 @@
 "use strict";
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
@@ -34,7 +35,6 @@ var __generator = (this && this.__generator) || function (thisArg, body) {
         if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
     }
 };
-var _this = this;
 exports.__esModule = true;
 var fetch = require("isomorphic-fetch");
 var log_1 = require("./log");
@@ -46,7 +46,7 @@ var log_1 = require("./log");
  * 如果需要自动化，参照fetch文档 https://developer.mozilla.org/zh-CN/docs/Web/API/Fetch_API/Using_Fetch
  * 在配置文件的fetchOption按RequestInit接口格式添加验证参数
  * */
-exports.fetchSwaggerJSONSchema = function (url, init) { return __awaiter(_this, void 0, void 0, function () {
+exports.fetchSwaggerJSONSchema = function (url, init) { return __awaiter(void 0, void 0, void 0, function () {
     var res, swaggerSchema;
     return __generator(this, function (_a) {
         switch (_a.label) {