+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.2.3](https://github.com/superwf/ts-gear/compare/v0.2.2...v0.2.3) (2019-10-09)
+
+
+### Features
+
+* add pathMatcher config ([53d2889](https://github.com/superwf/ts-gear/commit/53d2889a6d101cff110a934373c83650d9dea39d))
+* add pathMatcher function type ([07cbf8a](https://github.com/superwf/ts-gear/commit/07cbf8adabb8624542e77042958f4272dfc3945f))
+* 添加mockRequest请求参数打印 ([7962b9a](https://github.com/superwf/ts-gear/commit/7962b9a742d597ec07c5125d8a49b3cd049c30ec))
+
+
+### Bug Fixes
+
+* throw when use mockRequest in production ([40cdb17](https://github.com/superwf/ts-gear/commit/40cdb1739e2ada3923245530fde9078b09c67c48))
+* upgrade dependencies ([cfc8150](https://github.com/superwf/ts-gear/commit/cfc81502ce12de95e5f1f5ee09a5fcc2bf6ceca7))
+
 ### [0.2.2](https://github.com/superwf/ts-gear/compare/v0.2.1...v0.2.2) (2019-09-27)